fix: User innerWidth instead of outerWidth
When falling back for no native matchMedia support, window width was set
using `window.outerWidth`. This is the incorrect property to use, as
`outerWidth` includes window borders.

`window.innerWidth` only includes the scrollbar, which behaves the same
way as a `min-width` media query.

Note that this was mostly affecting IE9. Less than IE9 doesn't support
media queries, and greater than IE9 was never falling back to this
condition, as it natively supports matchMedia.